Add new instructions and operations to the virtual machine
- Implemented PUSH and POP instructions for stack operations.
- Added GETSP and SETSP instructions to manage the stack pointer.
- Introduced arithmetic operations: MUL, DIV, MOD, FMUL, FDIV.
- Added bitwise operations: AND, OR, XOR, NOT, SHL, SHR, SAR, NEG.
- Enhanced comparison operations: SLTS and EQ.
- Implemented 32-bit memory operations: LDM32, STM32, LDM32I, STM32I, and MOV32.
- Updated execution context to handle 32-bit values and added read_dword method.
- Modified CPU to use 32-bit registers instead of 16-bit.
- Added tests for new instructions and operations.
- Created demo programs showcasing string manipulation and table usage in fc-lang.

+++ crates/fc-vm/src/isa/operations/audio.rs
@@ -27,7 +27,7 @@ pub fn stop_music_opcode(ctx: &mut ExecutionContext) -> Result<(), VmFault> {
 pub fn play_sound(ctx: &mut ExecutionContext) -> Result<(), VmFault> {
     let freq = ctx.read_register_value()? as f32;
     let vol = ctx.read_register_value()? as f32 / 100.0;
-    let dur = ctx.read_register_value()?;
+    let dur = ctx.read_register_value()? as u16;
 
     debug!(
         "Playing square sound: freq={}, vol={}, dur={}",
@@ -63,7 +63,7 @@ pub fn play_sound_value(ctx: &mut ExecutionContext) -> Result<(), VmFault> {
 pub fn play_noise(ctx: &mut ExecutionContext) -> Result<(), VmFault> {
     let rate = ctx.read_register_value()? as f32;
     let vol = ctx.read_register_value()? as f32 / 100.0;
-    let dur = ctx.read_register_value()?;
+    let dur = ctx.read_register_value()? as u16;
 
     debug!("Playing noise: rate={}, vol={}, dur={}", rate, vol, dur);
     ctx.sound.noise = NoiseChannel {
